[BOT] refactor: extract Game utils
