+++ docs/Client/classes/GameUtils.md
@@ -1,0 +1,12 @@
+# GameUtils
+
+Defined in [`2006Scape Client/src/main/java/core/GameUtils.java`](../../2006Scape%20Client/src/main/java/core/GameUtils.java).
+
+Utility helpers extracted from the large [`Game`](Game.md) class.
+
+```java
+public final class GameUtils {
+    public static int random(float range)
+    public static String intToKOrMilLongName(int i)
+}
+```
